conditional image loading
Conditional Image Load checks to see the size of the div the image is to occupy. If the div is large enough, the script swaps odut the default image with a higher-resolution version. This way mobile loads aren't hung up loading large images and desktop users can see high-quality images.